Feat: Add my page in manager page, noti 조작
